# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OH Sansi API",
        "description": "Documentación de la API para OH Sansi.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        },
        {
            "url": "{APP_URL}",
            "description": "Servidor local",
            "variables": {
                "APP_URL": {
                    "default": "http://localhost:8000",
                    "description": "URL base de la aplicación"
                }
            }
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registro de usuario",
                "description": "Crea un nuevo usuario en el sistema",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "ci",
                                    "phone_number",
                                    "genre",
                                    "roles_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Maria"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Perez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@example.com"
                                    },
                                    "ci": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "71717717"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "enum": [
                                            "masculino",
                                            "femenino"
                                        ],
                                        "example": "femenino"
                                    },
                                    "roles_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User registration successfully"
                                        },
                                        "acces_token": {
                                            "type": "string",
                                            "example": "2|gfdsgsgfsgsgf..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error al crear el usuario"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Autenticar usuario y devolver token",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "maria@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "MP12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "2|gfdsgsgfsgsgf..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Elimina el token del usuario autenticado",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Logout exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}